requirements.txt
fastapi==0.110.1
uvicorn==0.25.0

# Environment & config
python-dotenv>=1.0.1
tzdata>=2024.2

# Database (MongoDB)
pymongo==4.5.0
motor==3.3.1

# Auth & Security
pyjwt>=2.10.1
python-jose>=3.3.0
bcrypt==4.1.3
passlib>=1.7.4
cryptography>=42.0.8
email-validator>=2.2.0

# Requests & OAuth
requests>=2.31.0
requests-oauthlib>=2.0.0
boto3>=1.34.129

# Forms & uploads
python-multipart>=0.0.9
